Move SteamVR code to dedicated source file
