+++ impl11/ddraw/Direct3DDevice.cpp
@@ -240,12 +240,14 @@ float s_XwaHudScale = 1.0f;
 //#include <assert.h>
 
 #include "FreePIE.h"
-#include <headers/openvr.h>
+#include "SteamVR.h"
+//#include <headers/openvr.h>
 #include "Matrices.h"
 
 #include "XWAObject.h"
 
 #include "..\shaders\shader_common.h"
+
 
 #define DBG_MAX_PRESENT_LOGS 0
 
@@ -342,7 +344,6 @@ const float DEFAULT_LENS_K3 =  0.0f;
 //const float DEFAULT_COCKPIT_PZ_THRESHOLD = 0.166f; // I used 0.13f for a long time until I jumped on a TIE-Interceptor
 const float DEFAULT_COCKPIT_PZ_THRESHOLD = 10.0f; // De-activated
 const int DEFAULT_SKYBOX_INDEX = 2;
-const bool DEFAULT_INTERLEAVED_REPROJECTION = false;
 const bool DEFAULT_BARREL_EFFECT_STATE = true;
 const bool DEFAULT_BARREL_EFFECT_STATE_STEAMVR = false; // SteamVR provides its own lens correction, only enable it if the user really wants it
 const float DEFAULT_BRIGHTNESS = 0.95f;
@@ -363,7 +364,6 @@ const float DEFAULT_MIN_POS_Y = -1.5f;
 const float DEFAULT_MAX_POS_Y =  1.5f;
 const float DEFAULT_MIN_POS_Z = -0.1f;
 const float DEFAULT_MAX_POS_Z =  2.5f;
-const bool DEFAULT_STEAMVR_POS_FROM_FREEPIE = false;
 const bool DEFAULT_RESHADE_ENABLED_STATE = false;
 const bool DEFAULT_BLOOM_ENABLED_STATE = false;
 // TODO: Make this toggleable later
@@ -442,27 +442,6 @@ typedef enum {
 VRModeEnum g_VRMode = VR_MODE_DIRECT_SBS;
 */
 
-// ***************************************************************
-// SteamVR HMD
-vr::IVRSystem *g_pHMD = NULL;
-vr::IVRCompositor *g_pVRCompositor = NULL;
-vr::IVRScreenshots *g_pVRScreenshots = NULL;
-vr::TrackedDevicePose_t g_rTrackedDevicePose;
-uint32_t g_steamVRWidth = 0, g_steamVRHeight = 0; // The resolution recommended by SteamVR is stored here
-bool g_bSteamVREnabled = false; // The user sets this flag to true to request support for SteamVR.
-bool g_bSteamVRInitialized = false; // The system will set this flag after SteamVR has been initialized
-bool g_bUseSteamVR = false; // The system will set this flag if the user requested SteamVR and SteamVR was initialized properly
-bool g_bInterleavedReprojection = DEFAULT_INTERLEAVED_REPROJECTION;
-bool g_bResetHeadCenter = true; // Reset the head center on startup
-bool g_bSteamVRDistortionEnabled = true;
-vr::HmdMatrix34_t g_EyeMatrixLeft, g_EyeMatrixRight;
-Matrix4 g_EyeMatrixLeftInv, g_EyeMatrixRightInv;
-Matrix4 g_projLeft, g_projRight;
-Matrix4 g_FullProjMatrixLeft, g_FullProjMatrixRight, g_viewMatrix;
-//float g_fMetricMult = DEFAULT_METRIC_MULT, 
-float g_fFrameTimeRemaining = 0.005f;
-int g_iSteamVR_Remaining_ms = 3, g_iSteamVR_VSync_ms = 11;
-
 bool g_bExternalHUDEnabled = false, g_bEdgeDetectorEnabled = false;
 
 // METRIC 3D RECONSTRUCTION
@@ -485,8 +464,7 @@ float g_fMinPositionX = DEFAULT_MIN_POS_X, g_fMaxPositionX = DEFAULT_MAX_POS_X;
 float g_fMinPositionY = DEFAULT_MIN_POS_Y, g_fMaxPositionY = DEFAULT_MAX_POS_Y;
 float g_fMinPositionZ = DEFAULT_MIN_POS_Z, g_fMaxPositionZ = DEFAULT_MAX_POS_Z;
 bool g_bStickyArrowKeys = false, g_bYawPitchFromMouseOverride = false;
-Vector3 g_headCenter; // The head's center: this value should be re-calibrated whenever we set the headset
-void projectSteamVR(float X, float Y, float Z, vr::EVREye eye, float &x, float &y, float &z);
+extern Vector3 g_headCenter; // The head's center: this value should be re-calibrated whenever we set the headset
 
 /* Vertices that will be used for the VertexBuffer. */
 D3DTLVERTEX *g_OrigVerts = NULL;
@@ -722,7 +700,6 @@ float g_fGUIElemsScale = DEFAULT_GLOBAL_SCALE; // Used to reduce the size of all
 int g_iFreePIESlot = DEFAULT_FREEPIE_SLOT;
 int g_iFreePIEControllerSlot = -1;
 bool g_bFixedGUI = DEFAULT_FIXED_GUI_STATE;
-bool g_bSteamVRPosFromFreePIE = DEFAULT_STEAMVR_POS_FROM_FREEPIE;
 bool g_bDirectSBSInitialized = false;
 //float g_fXWAScale = 1.0f; // This is the scale value as computed during Execute()
 D3D11_VIEWPORT g_nonVRViewport{};
@@ -4293,162 +4270,7 @@ next:
 	ReloadMaterials();
 }
 
-////////////////////////////////////////////////////////////////
-// SteamVR functions
-////////////////////////////////////////////////////////////////
-char *GetTrackedDeviceString(vr::TrackedDeviceIndex_t unDevice, vr::TrackedDeviceProperty prop, vr::TrackedPropertyError *peError = NULL)
-{
-	char *pchBuffer = NULL;
-	uint32_t unRequiredBufferLen = vr::VRSystem()->GetStringTrackedDeviceProperty(unDevice, prop, NULL, 0, peError);
-	if (unRequiredBufferLen == 0)
-		return pchBuffer;
 
-	pchBuffer = new char[unRequiredBufferLen];
-	unRequiredBufferLen = vr::VRSystem()->GetStringTrackedDeviceProperty(unDevice, prop, pchBuffer, unRequiredBufferLen, peError);
-	std::string sResult = pchBuffer;
-	return pchBuffer;
-}
-
-void DumpMatrix34(FILE *file, const vr::HmdMatrix34_t &m) {
-	if (file == NULL)
-		return;
-
-	for (int i = 0; i < 3; i++) {
-		for (int j = 0; j < 4; j++) {
-			fprintf(file, "%0.6f", m.m[i][j]);
-			if (j < 3)
-				fprintf(file, ", ");
-		}
-		fprintf(file, "\n");
-	}
-}
-
-void DumpMatrix44(FILE *file, const vr::HmdMatrix44_t &m) {
-	if (file == NULL)
-		return;
-
-	for (int i = 0; i < 4; i++) {
-		for (int j = 0; j < 4; j++) {
-			fprintf(file, "%0.6f", m.m[i][j]);
-			if (j < 3)
-				fprintf(file, ", ");
-		}
-		fprintf(file, "\n");
-	}
-}
-
-void DumpMatrix4(FILE *file, const Matrix4 &mat) {
-	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[0], mat[4], mat[8], mat[12]);
-	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[1], mat[5], mat[9], mat[13]);
-	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[2], mat[6], mat[10], mat[14]);
-	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[3], mat[7], mat[11], mat[15]);
-}
-
-void ShowMatrix4(const Matrix4 &mat, char *name) {
-	log_debug("[DBG] -----------------------------------------");
-	if (name != NULL)
-		log_debug("[DBG] %s", name);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[0], mat[4], mat[8], mat[12]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[1], mat[5], mat[9], mat[13]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[2], mat[6], mat[10], mat[14]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[3], mat[7], mat[11], mat[15]);
-	log_debug("[DBG] =========================================");
-}
-
-void ShowHmdMatrix34(const vr::HmdMatrix34_t &mat, char *name) {
-	log_debug("[DBG] -----------------------------------------");
-	if (name != NULL)
-		log_debug("[DBG] %s", name);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[0][0], mat.m[0][1], mat.m[0][2], mat.m[0][3]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[1][0], mat.m[1][1], mat.m[1][2], mat.m[1][3]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[2][0], mat.m[2][1], mat.m[2][2], mat.m[2][3]);
-	log_debug("[DBG] =========================================");
-}
-
-void ShowHmdMatrix44(const vr::HmdMatrix44_t &mat, char *name) {
-	log_debug("[DBG] -----------------------------------------");
-	if (name != NULL)
-		log_debug("[DBG] %s", name);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[0][0], mat.m[0][1], mat.m[0][2], mat.m[0][3]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[1][0], mat.m[1][1], mat.m[1][2], mat.m[1][3]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[2][0], mat.m[2][1], mat.m[2][2], mat.m[2][3]);
-	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[3][0], mat.m[3][1], mat.m[3][2], mat.m[3][3]);
-	log_debug("[DBG] =========================================");
-}
-
-void Matrix4toHmdMatrix44(const Matrix4 &m4, vr::HmdMatrix44_t &mat) {
-	mat.m[0][0] = m4[0];  mat.m[1][0] = m4[1];  mat.m[2][0] = m4[2];  mat.m[3][0] = m4[3];
-	mat.m[0][1] = m4[4];  mat.m[1][1] = m4[5];  mat.m[2][1] = m4[6];  mat.m[3][1] = m4[7];
-	mat.m[0][2] = m4[8];  mat.m[1][2] = m4[9];  mat.m[2][2] = m4[10]; mat.m[3][2] = m4[11];
-	mat.m[0][3] = m4[12]; mat.m[1][3] = m4[13]; mat.m[2][3] = m4[14]; mat.m[3][3] = m4[14];
-}
-
-Matrix4 HmdMatrix44toMatrix4(const vr::HmdMatrix44_t &mat) {
-	Matrix4 matrixObj(
-		mat.m[0][0], mat.m[1][0], mat.m[2][0], mat.m[3][0],
-		mat.m[0][1], mat.m[1][1], mat.m[2][1], mat.m[3][1],
-		mat.m[0][2], mat.m[1][2], mat.m[2][2], mat.m[3][2],
-		mat.m[0][3], mat.m[1][3], mat.m[2][3], mat.m[3][3]
-	);
-	return matrixObj;
-}
-
-Matrix4 HmdMatrix34toMatrix4(const vr::HmdMatrix34_t &mat) {
-	Matrix4 matrixObj(
-		mat.m[0][0], mat.m[1][0], mat.m[2][0], 0.0f,
-		mat.m[0][1], mat.m[1][1], mat.m[2][1], 0.0f,
-		mat.m[0][2], mat.m[1][2], mat.m[2][2], 0.0f,
-		mat.m[0][3], mat.m[1][3], mat.m[2][3], 1.0f
-	);
-	return matrixObj;
-}
-
-void ProcessSteamVREyeMatrices(vr::EVREye eye) {
-	if (g_pHMD == NULL) {
-		log_debug("[DBG] Cannot process SteamVR matrices because g_pHMD == NULL");
-		return;
-	}
-
-	vr::HmdMatrix34_t eyeMatrix = g_pHMD->GetEyeToHeadTransform(eye);
-	//ShowHmdMatrix34(eyeMatrix, "HmdMatrix34_t eyeMatrix");
-	if (eye == vr::EVREye::Eye_Left)
-		g_EyeMatrixLeft = eyeMatrix;
-	else
-		g_EyeMatrixRight = eyeMatrix;
-
-	Matrix4 matrixObj = HmdMatrix34toMatrix4(eyeMatrix);
-	/*
-	// Pimax matrix: 11.11 degrees on the Y axis and 6.64 IPD
-	Matrix4 matrixObj(
-		0.984808, 0.000000, 0.173648, -0.033236,
-		0.000000, 1.000000, 0.000000, 0.000000,
-		-0.173648, 0.000000, 0.984808, 0.000000,
-		0, 0, 0, 1);
-	matrixObj.transpose();
-	*/
-
-	/*
-	if (g_bInverseTranspose) {
-		log_debug("[DBG] Computing Inverse Transpose");
-		matrixObj.transpose();
-	}
-	*/
-	// Invert the matrix and store it
-	matrixObj.invertGeneral();
-	if (eye == vr::EVREye::Eye_Left)
-		g_EyeMatrixLeftInv = matrixObj;
-	else
-		g_EyeMatrixRightInv = matrixObj;
-}
-
-void ShowVector4(const Vector4 &X, char *name) {
-	if (name != NULL)
-		log_debug("[DBG] %s = %0.6f, %0.6f, %0.6f, %0.6f",
-			name, X[0], X[1], X[2], X[3]);
-	else
-		log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f",
-			X[0], X[1], X[2], X[3]);
-}
 
 void TestPimax() {
 	vr::HmdMatrix34_t eyeLeft;
@@ -4517,208 +4339,7 @@ void Test2DMesh() {
 	log_debug("[DBG] (%0.3f, %0.3f) --> (%0.3f, %0.3f)", P[0], P[1], Q[0], Q[1]);
 }
 
-bool InitSteamVR()
-{
-	/*
-	How to enable the null driver:
-	https://www.reddit.com/r/Vive/comments/6uo053/how_to_use_steamvr_tracked_devices_without_a_hmd/
 
-	steamvr.vrsettings file:
-	C:\Program Files (x86)\Steam\config
-
-	null driver config file:
-	C:\Program Files (x86)\Steam\steamapps\common\SteamVR\drivers\null\resources\settings
-	*/
-	char *strDriver = NULL;
-	char *strDisplay = NULL;
-	FILE *file = NULL;
-	//Matrix4 RollTest;
-	bool result = true;
-
-	int file_error = fopen_s(&file, "./steamvr_mat.txt", "wt");
-	log_debug("[DBG] Initializing SteamVR");
-	vr::EVRInitError eError = vr::VRInitError_None;
-	g_pHMD = vr::VR_Init(&eError, vr::VRApplication_Scene);
-	g_headCenter.set(0, 0, 0);
-
-	if (eError != vr::VRInitError_None)
-	{
-		g_pHMD = NULL;
-		log_debug("[DBG] Unable to init VR runtime: %s", vr::VR_GetVRInitErrorAsEnglishDescription(eError));
-		result = false;
-		goto out;
-	}
-	log_debug("[DBG] SteamVR Initialized");
-	g_pHMD->GetRecommendedRenderTargetSize(&g_steamVRWidth, &g_steamVRHeight);
-	log_debug("[DBG] Recommended steamVR width, height: %d, %d", g_steamVRWidth, g_steamVRHeight);
-
-	strDriver = GetTrackedDeviceString(vr::k_unTrackedDeviceIndex_Hmd, vr::Prop_TrackingSystemName_String);
-	if (strDriver) {
-		log_debug("[DBG] Driver: %s", strDriver);
-		strDisplay = GetTrackedDeviceString(vr::k_unTrackedDeviceIndex_Hmd, vr::Prop_SerialNumber_String);
-	}
-
-	if (strDisplay)
-		log_debug("[DBG] Display: %s", strDisplay);
-
-	g_pVRCompositor = vr::VRCompositor();
-	if (g_pVRCompositor == NULL)
-	{
-		log_debug("[DBG] SteamVR Compositor Initialization failed.");
-		result = false;
-		goto out;
-	}
-	log_debug("[DBG] SteamVR Compositor Initialized");
-
-	g_pVRCompositor->ForceInterleavedReprojectionOn(g_bInterleavedReprojection);
-	log_debug("[DBG] InterleavedReprojection: %d", g_bInterleavedReprojection);
-
-	g_pVRScreenshots = vr::VRScreenshots();
-	if (g_pVRScreenshots != NULL) {
-		log_debug("[DBG] SteamVR screenshot system enabled");
-	}
-
-	// Reset the seated pose
-	g_pHMD->ResetSeatedZeroPose();
-
-	// Pre-multiply and store the eye and projection matrices:
-	ProcessSteamVREyeMatrices(vr::EVREye::Eye_Left);
-	ProcessSteamVREyeMatrices(vr::EVREye::Eye_Right);
-
-	vr::HmdMatrix44_t projLeft = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Left, 0.001f, 100.0f);
-	vr::HmdMatrix44_t projRight = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Right, 0.001f, 100.0f);
-
-	g_projLeft  = HmdMatrix44toMatrix4(projLeft);
-	g_projRight = HmdMatrix44toMatrix4(projRight);
-
-	// Override all of the above with the Pimax matrices
-	//TestPimax();
-
-	g_FullProjMatrixLeft  = g_projLeft  * g_EyeMatrixLeftInv;
-	g_FullProjMatrixRight = g_projRight * g_EyeMatrixRightInv;
-
-	//Test2DMesh();
-
-	ShowMatrix4(g_EyeMatrixLeftInv, "g_EyeMatrixLeftInv");
-	ShowMatrix4(g_projLeft, "g_projLeft");
-
-	ShowMatrix4(g_EyeMatrixRightInv, "g_EyeMatrixRightInv");
-	ShowMatrix4(g_projRight, "g_projRight");
-	
-	// Initialize FreePIE if we're going to use it to read the position.
-	if (g_bSteamVRPosFromFreePIE)
-		InitFreePIE();
-
-	// Compute the FOV from the left eye vertical params. This FOV will be applied
-	// and saved in PrimarySurface::Flip()
-	float left, right, top, bottom;
-	g_pHMD->GetProjectionRaw(vr::EVREye::Eye_Left, &left, &right, &top, &bottom);
-	g_fVR_FOV = (atan(fabs(top)) + atan(fabs(bottom))) / DEG2RAD;
-
-	// Dump information about the view matrices
-	if (file_error == 0) {
-		Matrix4 eye, test;
-
-		if (strDriver != NULL)
-			fprintf(file, "Driver: %s\n", strDriver);
-		if (strDisplay != NULL)
-			fprintf(file, "Display: %s\n", strDisplay);
-		fprintf(file, "\n");
-
-		// Left Eye matrix
-		fprintf(file, "eyeLeft:\n");
-		DumpMatrix34(file, g_EyeMatrixLeft);
-		fprintf(file, "\n");
-
-		fprintf(file, "eyeLeftInv:\n");
-		DumpMatrix4(file, g_EyeMatrixLeftInv);
-		fprintf(file, "\n");
-
-		eye = HmdMatrix34toMatrix4(g_EyeMatrixLeft);
-		test = eye * g_EyeMatrixLeftInv;
-		fprintf(file, "Left Eye inverse test:\n");
-		DumpMatrix4(file, test);
-		fprintf(file, "\n");
-
-		// Right Eye matrix
-		fprintf(file, "eyeRight:\n");
-		DumpMatrix34(file, g_EyeMatrixRight);
-		fprintf(file, "\n");
-
-		fprintf(file, "eyeRightInv:\n");
-		DumpMatrix4(file, g_EyeMatrixRightInv);
-		fprintf(file, "\n");
-
-		eye = HmdMatrix34toMatrix4(g_EyeMatrixRight);
-		test = eye * g_EyeMatrixRightInv;
-		fprintf(file, "Right Eye inverse test:\n");
-		DumpMatrix4(file, test);
-		fprintf(file, "\n");
-
-		// Z_FAR was 50 for version 0.9.6, and I believe Z_Near was 0.5 (focal dist)
-		//vr::HmdMatrix44_t projLeft = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Left, DEFAULT_FOCAL_DIST, 50.0f);
-		//vr::HmdMatrix44_t projRight = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Right, DEFAULT_FOCAL_DIST, 50.0f);
-
-		fprintf(file, "projLeft:\n");
-		DumpMatrix44(file, projLeft);
-		fprintf(file, "\n");
-
-		fprintf(file, "projRight:\n");
-		DumpMatrix44(file, projRight);
-		fprintf(file, "\n");
-
-		g_pHMD->GetProjectionRaw(vr::EVREye::Eye_Left, &left, &right, &top, &bottom);
-		fprintf(file, "Raw data (Left eye):\n");
-		fprintf(file, "Left: %0.6f, Right: %0.6f, Top: %0.6f, Bottom: %0.6f\n",
-			left, right, top, bottom);
-
-		g_pHMD->GetProjectionRaw(vr::EVREye::Eye_Right, &left, &right, &top, &bottom);
-		fprintf(file, "Raw data (Right eye):\n");
-		fprintf(file, "Left: %0.6f, Right: %0.6f, Top: %0.6f, Bottom: %0.6f\n",
-			left, right, top, bottom);
-		fclose(file);
-	}
-
-out:
-	g_bSteamVRInitialized = result;
-
-	if (strDriver != NULL) {
-		delete[] strDriver;
-		strDriver = NULL;
-	}
-	if (strDisplay != NULL) {
-		delete[] strDisplay;
-		strDisplay = NULL;
-	}
-	if (file != NULL) {
-		fclose(file);
-		file = NULL;
-	}
-	return result;
-}
-
-void ShutDownSteamVR() {
-	// If the user sets g_bSteamVRPosFromFreePIE when XWA is starting, they may
-	// also reload the vrparams and then unset this flag. In that case, we won't
-	// shut FreePIE down. Well, this is a private hack, so I'm not going to care
-	// about that case.
-	if (g_bSteamVRPosFromFreePIE)
-		ShutdownFreePIE();
-
-	// We can't shut down SteamVR twice, we either shut it down here or in the cockpit look hook.
-	// It looks like the right order is to shut SteamVR down in the cockpit look hook
-	return;
-	//log_debug("[DBG] Not shutting down SteamVR, just returning...");
-	//return;
-	/*
-	log_debug("[DBG] Shutting down SteamVR...");
-	vr::VR_Shutdown();
-	g_pHMD = NULL;
-	g_pVRCompositor = NULL;
-	g_pVRScreenshots = NULL;
-	log_debug("[DBG] SteamVR shut down");
-	*/
-}
 
 bool InitDirectSBS()
 {
@@ -4800,40 +4421,7 @@ bool ShutDownDirectSBS() {
 	return true;
 }
 
-/*
-void ProcessVREvent(const vr::VREvent_t & event)
-{
-	switch (event.eventType)
-	{
-	case vr::VREvent_TrackedDeviceDeactivated:
-	{
-		log_debug("[DBG] Device %u detached.\n", event.trackedDeviceIndex);
-	}
-	break;
-	case vr::VREvent_TrackedDeviceUpdated:
-	{
-		log_debug("[DBG] Device %u updated.\n", event.trackedDeviceIndex);
-	}
-	break;
-	}
-}
-*/
 
-/*
-bool HandleVRInput()
-{
-	bool bRet = false;
-
-	// Process SteamVR events
-	vr::VREvent_t event;
-	while (g_pHMD->PollNextEvent(&event, sizeof(event)))
-	{
-		ProcessVREvent(event);
-	}
-
-	return bRet;
-}
-*/
 
 ////////////////////////////////////////////////////////////////
 // End of SteamVR functions
@@ -5319,31 +4907,6 @@ void DumpOrigVertices(FILE *file, int numVerts)
 	}
 }
 #endif
-
-/* 
- * In SteamVR, the coordinate system is as follows:
- * +x is right
- * +y is up
- * -z is forward
- * Distance is meters
- * This function is not used anymore and the projection is done in the vertex shader
- */
-void projectSteamVR(float X, float Y, float Z, vr::EVREye eye, float &x, float &y, float &z) {
-	Vector4 PX; // PY;
-
-	PX.set(X, -Y, -Z, 1.0f);
-	if (eye == vr::EVREye::Eye_Left) {
-		PX = g_FullProjMatrixLeft * PX;
-	} else {
-		PX = g_FullProjMatrixRight * PX;
-	}
-	// Project
-	PX /= PX[3];
-	// Convert to 2D
-	x =  PX[0];
-	y = -PX[1];
-	z =  PX[2];
-}
 
 /* Function to quickly enable/disable ZWrite. Currently only used for brackets */
 HRESULT Direct3DDevice::QuickSetZWriteEnabled(BOOL Enabled) {
