+++ impl11/ddraw/SteamVR.cpp
@@ -1,0 +1,523 @@
+#include <Windows.h>
+#include <stdio.h>
+#include "common.h"
+#include "utils.h"
+#include "SteamVR.h"
+#include <headers/openvr.h>
+#include "FreePIE.h"
+
+const float DEG2RAD = 3.141593f / 180.0f;
+vr::IVRSystem* g_pHMD = NULL;
+vr::IVRCompositor* g_pVRCompositor = NULL;
+vr::IVRScreenshots* g_pVRScreenshots = NULL;
+vr::TrackedDevicePose_t g_rTrackedDevicePose;
+uint32_t g_steamVRWidth = 0, g_steamVRHeight = 0; // The resolution recommended by SteamVR is stored here
+bool g_bSteamVREnabled = false; // The user sets this flag to true to request support for SteamVR.
+bool g_bSteamVRInitialized = false; // The system will set this flag after SteamVR has been initialized
+bool g_bUseSteamVR = false; // The system will set this flag if the user requested SteamVR and SteamVR was initialized properly
+bool g_bEnableSteamVR_QPC;
+const bool DEFAULT_INTERLEAVED_REPROJECTION = false;
+const bool DEFAULT_STEAMVR_POS_FROM_FREEPIE = false;
+bool g_bInterleavedReprojection = DEFAULT_INTERLEAVED_REPROJECTION;
+bool g_bResetHeadCenter = true; // Reset the head center on startup
+bool g_bSteamVRDistortionEnabled = true;
+Matrix4 g_EyeMatrixLeftInv, g_EyeMatrixRightInv;
+vr::HmdMatrix34_t g_EyeMatrixLeft, g_EyeMatrixRight;
+Matrix4 g_projLeft, g_projRight;
+Matrix4 g_FullProjMatrixLeft, g_FullProjMatrixRight, g_viewMatrix;
+//float g_fMetricMult = DEFAULT_METRIC_MULT, 
+float g_fFrameTimeRemaining = 0.005f;
+int g_iSteamVR_Remaining_ms = 3, g_iSteamVR_VSync_ms = 11;
+Vector3 g_headCenter;
+bool g_bSteamVRPosFromFreePIE = DEFAULT_STEAMVR_POS_FROM_FREEPIE;
+
+bool InitSteamVR();
+void ShutDownSteamVR();
+void ApplyFocalLength(float focal_length);
+bool UpdateXWAHackerFOV();
+void CycleFOVSetting();
+float ComputeRealVertFOV();
+float ComputeRealHorzFOV();
+float RealVertFOVToRawFocalLength(float real_FOV);
+
+
+
+bool InitSteamVR()
+{
+	/*
+	How to enable the null driver:
+	https://www.reddit.com/r/Vive/comments/6uo053/how_to_use_steamvr_tracked_devices_without_a_hmd/
+
+	steamvr.vrsettings file:
+	C:\Program Files (x86)\Steam\config
+
+	null driver config file:
+	C:\Program Files (x86)\Steam\steamapps\common\SteamVR\drivers\null\resources\settings
+	*/
+	char* strDriver = NULL;
+	char* strDisplay = NULL;
+	FILE* file = NULL;
+	//Matrix4 RollTest;
+	bool result = true;
+
+	int file_error = fopen_s(&file, "./steamvr_mat.txt", "wt");
+	log_debug("[DBG] Initializing SteamVR");
+	vr::EVRInitError eError = vr::VRInitError_None;
+	g_pHMD = vr::VR_Init(&eError, vr::VRApplication_Scene);
+	g_headCenter.set(0, 0, 0);
+
+	if (eError != vr::VRInitError_None)
+	{
+		g_pHMD = NULL;
+		log_debug("[DBG] Unable to init VR runtime: %s", vr::VR_GetVRInitErrorAsEnglishDescription(eError));
+		result = false;
+		goto out;
+	}
+	log_debug("[DBG] SteamVR Initialized");
+	g_pHMD->GetRecommendedRenderTargetSize(&g_steamVRWidth, &g_steamVRHeight);
+	log_debug("[DBG] Recommended steamVR width, height: %d, %d", g_steamVRWidth, g_steamVRHeight);
+
+	strDriver = GetTrackedDeviceString(vr::k_unTrackedDeviceIndex_Hmd, vr::Prop_TrackingSystemName_String);
+	if (strDriver) {
+		log_debug("[DBG] Driver: %s", strDriver);
+		strDisplay = GetTrackedDeviceString(vr::k_unTrackedDeviceIndex_Hmd, vr::Prop_SerialNumber_String);
+	}
+
+	if (strDisplay)
+		log_debug("[DBG] Display: %s", strDisplay);
+
+	g_pVRCompositor = vr::VRCompositor();
+	if (g_pVRCompositor == NULL)
+	{
+		log_debug("[DBG] SteamVR Compositor Initialization failed.");
+		result = false;
+		goto out;
+	}
+	log_debug("[DBG] SteamVR Compositor Initialized");
+
+	g_pVRCompositor->ForceInterleavedReprojectionOn(g_bInterleavedReprojection);
+	log_debug("[DBG] InterleavedReprojection: %d", g_bInterleavedReprojection);
+
+	g_pVRScreenshots = vr::VRScreenshots();
+	if (g_pVRScreenshots != NULL) {
+		log_debug("[DBG] SteamVR screenshot system enabled");
+	}
+
+	// Reset the seated pose
+	g_pHMD->ResetSeatedZeroPose();
+
+	// Pre-multiply and store the eye and projection matrices:
+	ProcessSteamVREyeMatrices(vr::EVREye::Eye_Left);
+	ProcessSteamVREyeMatrices(vr::EVREye::Eye_Right);
+
+	vr::HmdMatrix44_t projLeft = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Left, 0.001f, 100.0f);
+	vr::HmdMatrix44_t projRight = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Right, 0.001f, 100.0f);
+
+	g_projLeft = HmdMatrix44toMatrix4(projLeft);
+	g_projRight = HmdMatrix44toMatrix4(projRight);
+
+	// Override all of the above with the Pimax matrices
+	//TestPimax();
+
+	g_FullProjMatrixLeft = g_projLeft * g_EyeMatrixLeftInv;
+	g_FullProjMatrixRight = g_projRight * g_EyeMatrixRightInv;
+
+	//Test2DMesh();
+
+	ShowMatrix4(g_EyeMatrixLeftInv, "g_EyeMatrixLeftInv");
+	ShowMatrix4(g_projLeft, "g_projLeft");
+
+	ShowMatrix4(g_EyeMatrixRightInv, "g_EyeMatrixRightInv");
+	ShowMatrix4(g_projRight, "g_projRight");
+
+	// Initialize FreePIE if we're going to use it to read the position.
+	if (g_bSteamVRPosFromFreePIE)
+		InitFreePIE();
+
+	// Compute the FOV from the left eye vertical params. This FOV will be applied
+	// and saved in PrimarySurface::Flip()
+	float left, right, top, bottom;
+	g_pHMD->GetProjectionRaw(vr::EVREye::Eye_Left, &left, &right, &top, &bottom);
+	g_fVR_FOV = (atan(fabs(top)) + atan(fabs(bottom))) / DEG2RAD;
+
+	// Dump information about the view matrices
+	if (file_error == 0) {
+		Matrix4 eye, test;
+
+		if (strDriver != NULL)
+			fprintf(file, "Driver: %s\n", strDriver);
+		if (strDisplay != NULL)
+			fprintf(file, "Display: %s\n", strDisplay);
+		fprintf(file, "\n");
+
+		// Left Eye matrix
+		fprintf(file, "eyeLeft:\n");
+		DumpMatrix34(file, g_EyeMatrixLeft);
+		fprintf(file, "\n");
+
+		fprintf(file, "eyeLeftInv:\n");
+		DumpMatrix4(file, g_EyeMatrixLeftInv);
+		fprintf(file, "\n");
+
+		eye = HmdMatrix34toMatrix4(g_EyeMatrixLeft);
+		test = eye * g_EyeMatrixLeftInv;
+		fprintf(file, "Left Eye inverse test:\n");
+		DumpMatrix4(file, test);
+		fprintf(file, "\n");
+
+		// Right Eye matrix
+		fprintf(file, "eyeRight:\n");
+		DumpMatrix34(file, g_EyeMatrixRight);
+		fprintf(file, "\n");
+
+		fprintf(file, "eyeRightInv:\n");
+		DumpMatrix4(file, g_EyeMatrixRightInv);
+		fprintf(file, "\n");
+
+		eye = HmdMatrix34toMatrix4(g_EyeMatrixRight);
+		test = eye * g_EyeMatrixRightInv;
+		fprintf(file, "Right Eye inverse test:\n");
+		DumpMatrix4(file, test);
+		fprintf(file, "\n");
+
+		// Z_FAR was 50 for version 0.9.6, and I believe Z_Near was 0.5 (focal dist)
+		//vr::HmdMatrix44_t projLeft = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Left, DEFAULT_FOCAL_DIST, 50.0f);
+		//vr::HmdMatrix44_t projRight = g_pHMD->GetProjectionMatrix(vr::EVREye::Eye_Right, DEFAULT_FOCAL_DIST, 50.0f);
+
+		fprintf(file, "projLeft:\n");
+		DumpMatrix44(file, projLeft);
+		fprintf(file, "\n");
+
+		fprintf(file, "projRight:\n");
+		DumpMatrix44(file, projRight);
+		fprintf(file, "\n");
+
+		g_pHMD->GetProjectionRaw(vr::EVREye::Eye_Left, &left, &right, &top, &bottom);
+		fprintf(file, "Raw data (Left eye):\n");
+		fprintf(file, "Left: %0.6f, Right: %0.6f, Top: %0.6f, Bottom: %0.6f\n",
+			left, right, top, bottom);
+
+		g_pHMD->GetProjectionRaw(vr::EVREye::Eye_Right, &left, &right, &top, &bottom);
+		fprintf(file, "Raw data (Right eye):\n");
+		fprintf(file, "Left: %0.6f, Right: %0.6f, Top: %0.6f, Bottom: %0.6f\n",
+			left, right, top, bottom);
+		fclose(file);
+	}
+
+out:
+	g_bSteamVRInitialized = result;
+
+	if (strDriver != NULL) {
+		delete[] strDriver;
+		strDriver = NULL;
+	}
+	if (strDisplay != NULL) {
+		delete[] strDisplay;
+		strDisplay = NULL;
+	}
+	if (file != NULL) {
+		fclose(file);
+		file = NULL;
+	}
+	return result;
+}
+
+void ShutDownSteamVR() {
+	// If the user sets g_bSteamVRPosFromFreePIE when XWA is starting, they may
+	// also reload the vrparams and then unset this flag. In that case, we won't
+	// shut FreePIE down. Well, this is a private hack, so I'm not going to care
+	// about that case.
+	if (g_bSteamVRPosFromFreePIE)
+		ShutdownFreePIE();
+
+	// We can't shut down SteamVR twice, we either shut it down here or in the cockpit look hook.
+	// It looks like the right order is to shut SteamVR down in the cockpit look hook
+	return;
+	//log_debug("[DBG] Not shutting down SteamVR, just returning...");
+	//return;
+	/*
+	log_debug("[DBG] Shutting down SteamVR...");
+	vr::VR_Shutdown();
+	g_pHMD = NULL;
+	g_pVRCompositor = NULL;
+	g_pVRScreenshots = NULL;
+	log_debug("[DBG] SteamVR shut down");
+	*/
+}
+
+/*
+ * In SteamVR, the coordinate system is as follows:
+ * +x is right
+ * +y is up
+ * -z is forward
+ * Distance is meters
+ * This function is not used anymore and the projection is done in the vertex shader
+ */
+void projectSteamVR(float X, float Y, float Z, vr::EVREye eye, float& x, float& y, float& z) {
+	Vector4 PX; // PY;
+
+	PX.set(X, -Y, -Z, 1.0f);
+	if (eye == vr::EVREye::Eye_Left) {
+		PX = g_FullProjMatrixLeft * PX;
+	}
+	else {
+		PX = g_FullProjMatrixRight * PX;
+	}
+	// Project
+	PX /= PX[3];
+	// Convert to 2D
+	x = PX[0];
+	y = -PX[1];
+	z = PX[2];
+}
+
+/* Convenience function to call WaitGetPoses() */
+bool WaitGetPoses_QPC() {
+	static LARGE_INTEGER t0, t1, last_t, elapsed_us, elapsed_since_last_t_us, freq = { 0 };
+	uint64_t elapsed_ms, remaining_ms, waitgetposes_elapsed_ms;
+	bool result = false;
+
+	if (freq.QuadPart == 0) {
+		QueryPerformanceFrequency(&freq);
+		log_debug("[DBG] [QPF] freq: %lu", freq);
+	}
+	if (g_bEnableSteamVR_QPC) {
+		QueryPerformanceCounter(&t0);
+		// Compute the time elapsed since the previous last_t was taken
+		elapsed_since_last_t_us.QuadPart = t0.QuadPart - last_t.QuadPart;
+		elapsed_since_last_t_us.QuadPart *= 1000000;
+		elapsed_since_last_t_us.QuadPart /= freq.QuadPart;
+		//log_debug("[DBG] elapsed_since_last_t: %lu", elapsed_since_last_t_us.QuadPart);
+
+		// We want to call WaitGetPoses when we're about to reach a multiple of 11ms
+		// since the previous last_t. Say, we want to call it at either 8ms since last_t,
+		// 22-3ms = 19ms, 33-3ms = 30ms
+		elapsed_ms = elapsed_since_last_t_us.QuadPart / 1000;
+		// g_iSteamVR_VSync_ms default = 11
+		remaining_ms = g_iSteamVR_VSync_ms - (elapsed_ms % g_iSteamVR_VSync_ms);
+	}
+	else
+		remaining_ms = 0;
+
+	if (remaining_ms <= g_iSteamVR_Remaining_ms) {
+		// We need to call WaitGetPoses so that SteamVR gets the focus, otherwise we'll just get
+		// error 101 when calling VRCompositor->Submit
+		vr::EVRCompositorError error = g_pVRCompositor->WaitGetPoses(&g_rTrackedDevicePose,
+			0, NULL, 0);
+		if (g_bEnableSteamVR_QPC) {
+			QueryPerformanceCounter(&t1);
+			elapsed_us.QuadPart = t1.QuadPart - t0.QuadPart;
+			elapsed_us.QuadPart *= 1000000;
+			elapsed_us.QuadPart /= freq.QuadPart;
+			waitgetposes_elapsed_ms = elapsed_us.QuadPart / 1000;
+			//if (waitgetposes_elapsed_ms > remaining_ms)
+			//	log_debug("[DBG] waitgetposes_elapsed_ms: %d", waitgetposes_elapsed_ms);
+		}
+
+		//Sleep(2); // Using "20" here I get values like: elapsed_us: 20381 below, so that validates
+		// that elapsed_us.QuadPart is in microseconds.
+		result = true;
+		//log_debug("[DBG] WaitGetPoses");
+	}
+
+	if (g_bEnableSteamVR_QPC)
+		// Store this timestamp for the next frame
+		last_t = t1;
+	//log_debug("[DBG] elapsed_us: %lu", elapsed_us.QuadPart);
+	return result;
+}
+
+bool WaitGetPoses() {
+	vr::EVRCompositorError error = g_pVRCompositor->WaitGetPoses(&g_rTrackedDevicePose,
+		0, NULL, 0);
+	return true;
+}
+
+char* GetTrackedDeviceString(vr::TrackedDeviceIndex_t unDevice, vr::TrackedDeviceProperty prop, vr::TrackedPropertyError* peError)
+{
+	char* pchBuffer = NULL;
+	uint32_t unRequiredBufferLen = vr::VRSystem()->GetStringTrackedDeviceProperty(unDevice, prop, NULL, 0, peError);
+	if (unRequiredBufferLen == 0)
+		return pchBuffer;
+
+	pchBuffer = new char[unRequiredBufferLen];
+	unRequiredBufferLen = vr::VRSystem()->GetStringTrackedDeviceProperty(unDevice, prop, pchBuffer, unRequiredBufferLen, peError);
+	std::string sResult = pchBuffer;
+	return pchBuffer;
+}
+
+void DumpMatrix34(FILE* file, const vr::HmdMatrix34_t& m) {
+	if (file == NULL)
+		return;
+
+	for (int i = 0; i < 3; i++) {
+		for (int j = 0; j < 4; j++) {
+			fprintf(file, "%0.6f", m.m[i][j]);
+			if (j < 3)
+				fprintf(file, ", ");
+		}
+		fprintf(file, "\n");
+	}
+}
+
+void DumpMatrix44(FILE* file, const vr::HmdMatrix44_t& m) {
+	if (file == NULL)
+		return;
+
+	for (int i = 0; i < 4; i++) {
+		for (int j = 0; j < 4; j++) {
+			fprintf(file, "%0.6f", m.m[i][j]);
+			if (j < 3)
+				fprintf(file, ", ");
+		}
+		fprintf(file, "\n");
+	}
+}
+
+void DumpMatrix4(FILE* file, const Matrix4& mat) {
+	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[0], mat[4], mat[8], mat[12]);
+	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[1], mat[5], mat[9], mat[13]);
+	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[2], mat[6], mat[10], mat[14]);
+	fprintf(file, "%0.6f, %0.6f, %0.6f, %0.6f\n", mat[3], mat[7], mat[11], mat[15]);
+}
+
+void ShowMatrix4(const Matrix4& mat, char* name) {
+	log_debug("[DBG] -----------------------------------------");
+	if (name != NULL)
+		log_debug("[DBG] %s", name);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[0], mat[4], mat[8], mat[12]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[1], mat[5], mat[9], mat[13]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[2], mat[6], mat[10], mat[14]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat[3], mat[7], mat[11], mat[15]);
+	log_debug("[DBG] =========================================");
+}
+
+void ShowHmdMatrix34(const vr::HmdMatrix34_t& mat, char* name) {
+	log_debug("[DBG] -----------------------------------------");
+	if (name != NULL)
+		log_debug("[DBG] %s", name);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[0][0], mat.m[0][1], mat.m[0][2], mat.m[0][3]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[1][0], mat.m[1][1], mat.m[1][2], mat.m[1][3]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[2][0], mat.m[2][1], mat.m[2][2], mat.m[2][3]);
+	log_debug("[DBG] =========================================");
+}
+
+void ShowHmdMatrix44(const vr::HmdMatrix44_t& mat, char* name) {
+	log_debug("[DBG] -----------------------------------------");
+	if (name != NULL)
+		log_debug("[DBG] %s", name);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[0][0], mat.m[0][1], mat.m[0][2], mat.m[0][3]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[1][0], mat.m[1][1], mat.m[1][2], mat.m[1][3]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[2][0], mat.m[2][1], mat.m[2][2], mat.m[2][3]);
+	log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f", mat.m[3][0], mat.m[3][1], mat.m[3][2], mat.m[3][3]);
+	log_debug("[DBG] =========================================");
+}
+
+void Matrix4toHmdMatrix44(const Matrix4& m4, vr::HmdMatrix44_t& mat) {
+	mat.m[0][0] = m4[0];  mat.m[1][0] = m4[1];  mat.m[2][0] = m4[2];  mat.m[3][0] = m4[3];
+	mat.m[0][1] = m4[4];  mat.m[1][1] = m4[5];  mat.m[2][1] = m4[6];  mat.m[3][1] = m4[7];
+	mat.m[0][2] = m4[8];  mat.m[1][2] = m4[9];  mat.m[2][2] = m4[10]; mat.m[3][2] = m4[11];
+	mat.m[0][3] = m4[12]; mat.m[1][3] = m4[13]; mat.m[2][3] = m4[14]; mat.m[3][3] = m4[14];
+}
+
+Matrix4 HmdMatrix44toMatrix4(const vr::HmdMatrix44_t& mat) {
+	Matrix4 matrixObj(
+		mat.m[0][0], mat.m[1][0], mat.m[2][0], mat.m[3][0],
+		mat.m[0][1], mat.m[1][1], mat.m[2][1], mat.m[3][1],
+		mat.m[0][2], mat.m[1][2], mat.m[2][2], mat.m[3][2],
+		mat.m[0][3], mat.m[1][3], mat.m[2][3], mat.m[3][3]
+	);
+	return matrixObj;
+}
+
+Matrix4 HmdMatrix34toMatrix4(const vr::HmdMatrix34_t& mat) {
+	Matrix4 matrixObj(
+		mat.m[0][0], mat.m[1][0], mat.m[2][0], 0.0f,
+		mat.m[0][1], mat.m[1][1], mat.m[2][1], 0.0f,
+		mat.m[0][2], mat.m[1][2], mat.m[2][2], 0.0f,
+		mat.m[0][3], mat.m[1][3], mat.m[2][3], 1.0f
+	);
+	return matrixObj;
+}
+
+void ProcessSteamVREyeMatrices(vr::EVREye eye) {
+	if (g_pHMD == NULL) {
+		log_debug("[DBG] Cannot process SteamVR matrices because g_pHMD == NULL");
+		return;
+	}
+
+	vr::HmdMatrix34_t eyeMatrix = g_pHMD->GetEyeToHeadTransform(eye);
+	//ShowHmdMatrix34(eyeMatrix, "HmdMatrix34_t eyeMatrix");
+	if (eye == vr::EVREye::Eye_Left)
+		g_EyeMatrixLeft = eyeMatrix;
+	else
+		g_EyeMatrixRight = eyeMatrix;
+
+	Matrix4 matrixObj = HmdMatrix34toMatrix4(eyeMatrix);
+	/*
+	// Pimax matrix: 11.11 degrees on the Y axis and 6.64 IPD
+	Matrix4 matrixObj(
+		0.984808, 0.000000, 0.173648, -0.033236,
+		0.000000, 1.000000, 0.000000, 0.000000,
+		-0.173648, 0.000000, 0.984808, 0.000000,
+		0, 0, 0, 1);
+	matrixObj.transpose();
+	*/
+
+	/*
+	if (g_bInverseTranspose) {
+		log_debug("[DBG] Computing Inverse Transpose");
+		matrixObj.transpose();
+	}
+	*/
+	// Invert the matrix and store it
+	matrixObj.invertGeneral();
+	if (eye == vr::EVREye::Eye_Left)
+		g_EyeMatrixLeftInv = matrixObj;
+	else
+		g_EyeMatrixRightInv = matrixObj;
+}
+
+void ShowVector4(const Vector4& X, char* name) {
+	if (name != NULL)
+		log_debug("[DBG] %s = %0.6f, %0.6f, %0.6f, %0.6f",
+			name, X[0], X[1], X[2], X[3]);
+	else
+		log_debug("[DBG] %0.6f, %0.6f, %0.6f, %0.6f",
+			X[0], X[1], X[2], X[3]);
+}
+
+/*
+void ProcessVREvent(const vr::VREvent_t & event)
+{
+	switch (event.eventType)
+	{
+	case vr::VREvent_TrackedDeviceDeactivated:
+	{
+		log_debug("[DBG] Device %u detached.\n", event.trackedDeviceIndex);
+	}
+	break;
+	case vr::VREvent_TrackedDeviceUpdated:
+	{
+		log_debug("[DBG] Device %u updated.\n", event.trackedDeviceIndex);
+	}
+	break;
+	}
+}
+*/
+
+/*
+bool HandleVRInput()
+{
+	bool bRet = false;
+
+	// Process SteamVR events
+	vr::VREvent_t event;
+	while (g_pHMD->PollNextEvent(&event, sizeof(event)))
+	{
+		ProcessVREvent(event);
+	}
+
+	return bRet;
+}
+*/
